% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_est_moe.R
\name{arrange_est_moe}
\alias{arrange_est_moe}
\title{Arrange estimate and MOE}
\usage{
arrange_est_moe(.data, omit = 1)
}
\arguments{
\item{.data}{A tbl.}

\item{omit}{An integer or vector of integers corresponding to the position of the column(s) that shouldn't be rearranged. Intended use case is where first 1-2 columns are characters such as names and counties. Defaults to 1.}
}
\value{
Returns a tbl, making it useful in dplyr chaining
}
\description{
Neatly arrange estimate and margin of error columns, such that each estimate column is followed by a column with the corresponding margin of error.
}
