% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_acs_table.R
\name{calc_acs_table}
\alias{calc_acs_table}
\title{Calculate ACS table}
\usage{
calc_acs_table(groups, denom, table)
}
\arguments{
\item{groups}{A named list. Each item should be a numeric vector of column positions. Names will be used as aggregation terms}

\item{denom}{A single-column ACS object, or one column of a larger ACS object, to be used as the denominator}

\item{table}{An ACS object where \code{groups} columns are located}
}
\value{
Returns an \code{\link[acs]{acs-class}} object, as is standard from the \code{acs} package.
}
\description{
Calculates aggregated sums and percentages for groups of ACS variables
}
\examples{
# Calculate the share of CT population below poverty and below 2x poverty (numerators), out of all residents for whom poverty status is determined (denominator)
poverty_table <- acs.fetch(geography = geo.make(state = 9), endyear = 2015, table.number = "C17002", col.names = "pretty")
pov_status_determined <- poverty[, 1]
pov_groups <- list(
  below_poverty = 2:3,
  low_income = 2:7
)
share_in_poverty <- calc_acs_table(pov_groups, pov_status_determined, poverty_table)
}
