% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_town_names.R
\name{get_town_names}
\alias{get_town_names}
\title{Get list of town names}
\usage{
get_town_names(state = 9, year = 2015)
}
\arguments{
\item{state}{Either a number corresponding to the state's FIPS code, or a string with the state's two-letter abbreviation or full name. Defaults to 09, FIPS code for Connecticut}

\item{year}{Endyear of ACS estimates as a four-digit integer. Defaults to 2015, most recent ACS estimates at time of writing}
}
\value{
Returns a data frame
}
\description{
Gets a tbl of all towns and counties in a given state
}
\examples{
ct_towns <- get_town_names(state = 9)
}
