% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfd_tracts.R
\docType{data}
\name{hfd_tracts}
\alias{hfd_tracts}
\title{Hartford, CT neighborhoods by tract}
\format{A list of 16 tbls
\describe{
  \item{tract}{FIPS code for a Census tract}
}}
\usage{
hfd_tracts
}
\description{
A named list of tbls, giving FIPS codes of Census tracts making up each neighborhood.
}
\keyword{datasets}
