% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_regional_geo.R
\name{make_regional_geo}
\alias{make_regional_geo}
\title{Create regional geography}
\usage{
make_regional_geo(town_list, name = "Aggregate", town_lookup, state = 9)
}
\arguments{
\item{town_list}{A character vector of the names or FIPS codes of towns in the region}

\item{name}{Name of the region. Defaults to "Aggregate"}

\item{town_lookup}{A data.frame or tbl. Must have columns \code{town} and \code{county}, as generated by \code{\link{get_town_names}}}

\item{state}{Either a number corresponding to the state's FIPS code, or a string with the state's two-letter abbreviation or full name. Defaults to 09, FIPS code for Connecticut}
}
\value{
Returns a \code{\link[acs]{geo.set}} object.
}
\description{
Merge towns into an ACS geo set
}
\examples{
inner_ring <- c("Hamden", "East Haven", "West Haven")
ct_towns <- get_town_names(state = 9, year = 2015)
ir_geo <- make_regional_geo <- function(inner_ring, name = "Inner Ring", town_lookup = ct_towns, state = 9)

}
\seealso{
\code{\link{get_town_names}}, which returns a dataframe of towns and countied formatted for \code{town_lookup}
}
