% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_years_for_trend.R
\name{make_years_for_trend}
\alias{make_years_for_trend}
\title{Get years available in Census API}
\usage{
make_years_for_trend(start_year, end_year)
}
\arguments{
\item{start_year}{A four-digit integer for the first year in the desired time period}

\item{end_year}{A four-digit integer for the last year in the desired time period}
}
\value{
Returns a tbl
}
\description{
Convenience function for splitting a span of years based on availability in the Census ACS API
}
\details{
The Census ACS API is currently only available for 2012 onward. This makes it tricky to pull data for a longer period, such as 2010 to 2015. The \code{type} column in the returned dataframe is useful for iterating over years that must be downloaded as csv files, versus those available from the API.
}
\examples{
years <- make_years_for_trend(2010, 2015)
}
