% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhv_neighborhoods.R
\docType{data}
\name{nhv_neighborhoods}
\alias{nhv_neighborhoods}
\title{New Haven, CT neighborhoods by tract and block group}
\format{A list of length 23
\describe{
  \item{tract}{FIPS code for a Census tract}
  \item{block.group}{numeric vector of FIPS codes for block groups}
}}
\usage{
nhv_neighborhoods
}
\description{
A named, nested list of lists (of lists, in some cases). Each outer list item represents one neighborhood.
}
\keyword{datasets}
