% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhv_tracts.R
\docType{data}
\name{nhv_tracts}
\alias{nhv_tracts}
\title{New Haven, CT neighborhoods by tract only}
\format{A list of length 22
\describe{
  \item{tract}{FIPS code for a Census tract}
}}
\usage{
nhv_tracts
}
\description{
A named, nested list of lists (of lists, in some cases). Each outer list item represents one neighborhood.
}
\keyword{datasets}
