% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\docType{data}
\name{regions}
\alias{regions}
\title{Regions of towns in Connecticut}
\format{A list of length 8}
\usage{
regions
}
\description{
A named list of character vectors giving the names or FIPS codes of towns in each region.
}
\keyword{datasets}
